+++ packages/elements/layer-manager/__tests__/index.js
@@ -14,8 +14,13 @@ function assertEqual(actual, expected) {
 }
 
 // NOTE: shim avoids noise in test logs
-// eslint-disable-next-line react/prop-types, no-unused-vars
-const Shim = ({ stackIndex, stackTotal, ...props }) => <span {...props} />;
+const Shim = ({
+  // eslint-disable-next-line react/prop-types
+  stackIndex, // eslint-disable-next-line react/prop-types
+  stackTotal, // eslint-disable-next-line react/prop-types
+  blockChildGatewayRender, // eslint-disable-next-line react/prop-types
+  ...props
+}) => <span {...props} />;
 
 describe(name, () => {
   it('should render Gateway in GatewayDest', () => {