+++ packages/fabric/editor-bitbucket-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-bitbucket-transformer
 
+## 2.5.5
+
 ## 2.5.4
 
 ## 2.5.3