+++ packages/fabric/editor-common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-common
 
+## 6.1.5
+
 ## 6.1.4
 
 ## 6.1.3